MotionProfile code (FHE)
https://www.ctr-electronics.com/downloads/pdf/Talon%20SRX%20Motion%20Profile%20Reference%20Manual.pdf

The intent of the motion control features of Talon is allow the robot
controller to generate and
stream the trajectory points into the Talon SRX while the Talon honors
the currently-selected
control mode. This means that the Talon SRX can be in any control-mode
(PercentVoltage for
example) while motion profile buffering occurs. When the robot
controller has detected that
enough of the trajectory points have been funneled into the Talon, the
robot controller can then
select the motion profile control mode, and enable the MPE by setting a
‘1’,